Modified pkg structure
